TP 64850
Somehow missed a commit of this domain controller.

Did some final work on this.  I had to refactor an overriding mail builder inorder to get this to work.  There was previously an effort to make the ability to send mail configurable across multiple organizations, it failed.  The solution will currently allow this as long as all organizations email servers use SMTP with the same level of security.  It wouldn't be too much more work to get it be be my configurable, but I don't believe it's worth the effort at this time.

git-svn-id: svn+ssh://cvs1/svn/listen/trunk@2912 4692729f-32a0-46b7-9832-1bf53c4881b5