TP 66777 - Changed up the FMFM service a fair amount.  It now only creates 1 cdr, where before it was creating 2 for every outbound call.  I've also added some logic to get the call duration in the CDR.
The bigger change is that I've made the FMFM service a subroutine of the Listen Merge app now.  This service is accessed by the same service id that is used by the conferencing app and is accessed by the variable LISTEN_SERVICEID set within the Listen Merge app on the Start page.

git-svn-id: svn+ssh://cvs1/svn/listen/trunk@3320 4692729f-32a0-46b7-9832-1bf53c4881b5